refactor: fix some issue with -W gcc flags, improve file structure for lib

+++ src/cmd/delete.c
@@ -3,7 +3,8 @@
 void delete(struct db *db, char *arg) {
   while (*arg == ' ')
     arg++;
-  int idx = atoi(arg);
+  // int idx = atoi(arg);
+  size_t idx = atoi(arg);
 
   if (idx >= db->count) {
     puts("invalid index");
